+++ RedPandaIDE/editor.cpp
@@ -61,6 +61,7 @@ Editor::Editor(QWidget *parent):
 {
 }
 
+
 Editor::Editor(QWidget *parent, const QString& filename,
                   const QByteArray& encoding,
                   Project* pProject, bool isNew,